requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.35
langchain==0.3.2
pydantic==2.9.0
python-dotenv==1.0.1
gradio==5.5.0
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.1
sqlalchemy==2.0.34
httpx==0.27.2
jinja2==3.1.4